pyproject.toml
[project]
name = "dreame-mocker"
version = "0.1.0"
description = "Python client that mocks the Dreame mobile app, enabling Home Assistant automation for the X50 Ultra Complete"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "gmqtt>=0.6.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pycryptodome>=3.20.0",
    "tenacity>=8.0.0",
]

[project.scripts]
dreame-mocker = "dreame_mocker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.408",
    "python-dotenv>=1.0.0",
]
